move + refactor lsp
