+++ src/lsp/__init__.py
@@ -1,0 +1,3 @@
+from lsp.main import run
+
+__all__ = ["run"]
