+++ src/lsp/main.py
@@ -1,0 +1,5 @@
+from lsp.lsp_runner import start_server
+
+
+def run():
+    start_server()
