--- src/vscode-atopile/bundled/tool/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
